Fix checking what needs updating in ansible.
When we update a collection or other resource, we make a check to see if
the values we want to set are already present.  This is done via
`set(reosurce.items())`, which fails if any of the values in the
resource dictionary are unhashable, such as the access flags list.  This
only checks the values that could possibly be relevant, and, if the
comparison still throws a `TypeError`, updates the resource anyway.