Fix order of module resolution for plugins
Prior to this change, each plugin was overwriting the module
resolution list and hard-coding the girder/node_modules dir
as the first element. This order was incorrect; we actually
want it to look inside its own node_modules first. After this change,
the order that will be searched for both modules and loaders is:

1. Inside the plugin's own `node_modules` dir.
2. Inside girder's `node_modules` dir.