Fix rest requests honoring `type`.
An unintentional breaking change was introduced in PR #2273.  Before, our ajax defaults (rightly or wrongly) preferred `type` over `method`, and set a default `type`.  jQuery's ajax call prefers `method`, so if a caller used either `type` *or* `method`, we would honor the caller's value.  Now, since we set a default `method`, if a caller uses `method`, it is ignored.

Further, we had previous set the dataType default to 'json'.  PR #2273 removed this, which then allows jQuery to use "Intelligent Guess".  Since this is also a breaking change, I've reverted it.