Initial FUSE plugin.
Assuming that the OS has fuse support, if this plugin is enabled AND there is a config entry in the [server_fuse] section of "path = <directory>", a FUSE will be mounted at that directory giving access to all files in Girder.

Additional mounts can be made with specific user access.

Note that if Girder is SIGKILLed with open file handles on the FUSE, it may not be possible to fully clean up the open file system, and defunct processes may linger.

Possible future work:

- Allow creating a mount with a specific list of accessible items or files.  This could be used by girder_worker instead of downloading when there is a shared file system.
- Allow writing.  One way would be to only create folders or items in Girder when a file is written, as that way it would be certain which things are folders versus items.  Or perhaps we could convert a folder to an item or vice versa if the only thing in it was folders or files.
- Expose a user-authenticated mount point for specific file browsing.