Store the optional "reference" in the item_tasks spec
This adds a new optional field in the item_tasks spec to store
extra information about the parameter.  Specifically, this is
needed to save the reference attribute that is defined in the
cli_ctk spec.  We code codify this more explicitly on the item
tasks spec, but the definition of this attribute is not terribly
clear and seems to be very task specific.