Remove the test for adding the parameters references
The addition of the reference attribute in ctk-cli is included in
an unmerged PR:

https://github.com/commontk/ctk-cli/pull/8

HistomicsTK relies on a fork of ctk-cli for this addition.  Rather
than remove the feature all together, I have removed the test so
that item_tasks doesn't rely on a fork.  Instead, we will install
the fork for HistomicsTK deployments.