Simplify python/virtualenv installation process
Prior to this, quite a bit of complex (and outright incorrect) logic
was implemented in order to manage the correct python/pip/virtualenv
to use in any given context. This commit removes almost all of the
complexity and adds the requirement that a user must bring their own
virtualenv if they want to install against a different Python than the
one Ansible is running under.