Add a registry for loading and querying plugins
This adds a module meant to replace the functional contents of
utilities/plugin_utilities.py for the girder 3.0 migration.  This module
is not yet used to maintain backwards compatibility.  I have added it as
an independent PR to collect feedback before moving on with the larger
effort to replace core girder code with this new module, which will
require backwards incompatible changes and migration of existing
plugins.