Add integration tests.
Gather and fix coverage.

The python system test has been moved to the integration test group since it requires building the version json record.

Hardened one test to pass in Python 3.5.

Packaging tests are now always run; we can switch back to only running on master and branches if we decide it is too slow.

With this commit, all tests are processed in a CircleCI version 2 workflow.  CircleCI version 2 has two executor environments, docker and machine, where the docker starts faster but is more limited.  The new workflow is divided into eight build steps:
- 1. docker executor for python2 tests
- 2. machine executor for python2 tests that can't run in docker (ones that test mongo sharding, for instance)
- 3.  docker executor that collects coverage from the python2 tests (from steps 1 and 2)
- 4. docker executor to configure python3 environment
- 5. docker executor for web client tests in a python3 environment (requires step 4)
- 6. docker executor for python3 tests (requires step 4)
- 7. machine executor for python3 tests
- 8.  docker executor that collects coverage from the python3 and web client tests (from steps 5, 6, and 7)

The total test time seems to be around 10 minutes (the CircleCI version 1 tests took around 14 to 18 minutes).  This requires more parallel slots for maximum speed, but since it is faster, we should gain in through-put.

Future work:
- reduce number of files send between workflow steps.  This would speed things up (perhaps as much as 40 seconds).
- refactor how we compute system version information so we can move the system test to the docker environment.
- refactor the .circleci/config.yml to reduce duplication
- add or convert tests to work with Python 3.6, too.  Current python tests are version 3.4 in the docker executor and 3.5 in the machine executor.
- improve access to build artifacts to make it easier to analyze errors.