Set a maximum height for modal dialogs.
Currently, if a modal dialog is tall, the webpage as a whole gets a
scrollbar and the dialog extends past the bottom.  I think it looks
better to have the dialog have a scroll bar (since the page may already
be scrolled).  The bootstrap css has different margins depending on
window size; this makes one policy that is consistent.