Swagger doesn't recognize application/zip as binary data.
Swagger injects BOM markers to non-binary data when it buffers it for
downloading.  Since Swagger 2.x only recognizes images, pdfs, and
application/octet-stream as binary, any other content dispositions can
be corrupted by this process.  Change the value emitted for the
`produces` field for Swagger so that application/zip is converted to
application/octet-stream.