Add model-layer support for virtual folders
These folders cannot contain subfolders or items in the normal
way, but can be used to display an "alternate" view of the
hierarchy via a custom database query to control what folders
and items are listed as their children. Those folders and items
must live somewhere else within the hierarchy.