Add a findWithPermissions method.
For access controlled models and models that use the acl-mixin, add a
`findWithPermissions` method that uses mongo to filter by permissions
either by extra query parameters (for ac models) or via an aggregate
pipeline (for acl-mixin models).  For models that use the acl-mixin and
get permissions from a model that itself uses the acl-mixin or use
`attachedToType` rather than `resourceColl`, this falls back to using
`filterResultsByPermission`.  If an aggregate pipeline is used, augment
the resulting CommandCursor with a `count` method that uses a similar
pipeline to get the count (but only on demand).

The where clauses used for groups are different than for other models.

For rest endpoints, support returning generators, Mongo Cursors, and
Mongo Command Cursors.  For each of these, the result is coerced to a
list as part of the standard endpoint handling.  For Cursors and Command
Cursors with a count method, an 'X-Total-Count' header to the response.

If the Mongo server is less than version 3.4, this falls back to
`filterResultsByPermission`.

Things that could be done:

* For models that use the acl-mixin and refer to a model that uses an
  acl-mixin, we could do multiple joins ($lookup) instead of falling
  back to filtering in python.

* Contribute to the mongo proxy upstream so that we can use it directly.

* When `findWithPermissions` is called without a user and asking for
  more than READ permission, a special query is constructed that returns
  an empty result set.  This seems hacky, but results in a consistent
  result type (a Cursor or CommandCursor).  Is there a better way to do
  this?

Areas that need testing:

* The mongo proxy with a command cursor.

* We may want to test with Mongo version 3.4 (and Python 2 with Mongo
  >=3.4).