Remove symlinked file to fix pip upgrade
Girder's custom `InstallWithOptions` setup.py command attempts to merge
the new package into an existing installation to avoid overriding
user-installed plugins and built static assets.  To do this, it uses
`distutils.dir_util.copy_tree` which doesn't appear to handle symlinks
correctly (or as we expect it would).  When upgrading an install, it
throws an error:

```
  FileExistsError: [Errno 17] File exists: 'setup.yml' -> \
    'girder/plugins/treeview/plugin_tests/treeviewWidgetSpec.yml'
```

Because the symlink wasn't necessary anyway, this removes it, but there
remain bugs in the implementation:

1. `plugin_tests` should be excluded from the manifest, but aren't.
2. Relative symlinks aren't being tracked and removed prior to upgrade
   by setuptools.  This could be a bug in distutils or more likely in
   how we are using it.