Add a pytest marker to enable plugins
This commit makes it possible to write pytest style tests for girder
plugins.  It is based on the interface that will be provided by pip
installable plugins in girder 3.  The new marker is incompatible with
the "testPlugin" marker, which mocks the normal plugin load mechanism.
This could in theory be fixed, but it isn't clear there is a use case
for enabling both "test plugins" and "real plugins" in a single test.