Garbage collect before unmounting fuse.
If a file has been placed in the queue for garbage collection, it can
still be open until python actually does the garbage collection.  If the
file is on a fuse mount, fuse can't unmount the file system due to the
open file.  By running `gc.collect()` prior to the unmount call, files
that should be closed don't prevent the unmounting.

Prior to this change, there were some wait loops in the tests to try to
clean up that still sometimes could fail due to the file being open and
the mount not closing.