Add --upgrade-strategy eager to all uses of pip install -U
The default behavior of `pip install -U` has changed as of version 10.
The default only-if-needed strategy only upgrades direct dependencies of
the packages being installed.  This is likely to break automated
upgrades such as ansible and CI deployments that use a previously
generated virtual environment.

Both upgrade strategies can lead to invalid virtual environment states
because resolving version conflicts is not something pip claims to
solve.  One could argue that the new behavior is superior because it
makes minimal changes to the environment; however, my guess is that the
eager strategy will cause less confusion and is identical to what pip
had done prior to version 10.

See also:

https://github.com/pypa/pip/issues/5274