Add a --verbose flag to girder_client
This commit creates a root logger instance inside the girder client cli
and configures logging on the urllib3 logger instance to output to
stderr.  The verbose flag can be provided multiple times for additional
verbosity.  The highest level (`girder-client -vvv`) will show all http
requests made.  From what I can tell, lower levels will only output
something in the event of an error.

This generates a root logger instance, which is appropriate because it
is only for a CLI usage.  If we ever migrate the print statements in the
girder-client library to loggers, this flag will control it as well.