Improve the access widget.
If the access widget does not show the access type, it would store
`level: Nan` in the access records.  This changes it to store `0`
instead.

If enter was typed in the search field of the access widget, the UI
would switch to a results page, losing any changes in the access widget.
This adds an option to the search field widget to not jump to a results
page.