Improve the duration tooltips on the job time graphs.
This fixes an issue with times that were exactly 1 minute or 1 hour or
were greater than or equal to a day.  Durations are shown in different
ways depending if they are less than a minute, between a minute and an
hour, or longer than an hour.

Before this change, durations were displayed, e.g., 01.234s, 01:23.456s,
01:23:45.678s, where there was often a needless leading zero and always
a trailing 's' implying seconds.  Further, values greater than 24 hours
would show the duration modulo 24 hours.  Due to the wrong inequality
being used, a duration of exactly 1 minute would show as 00.000s, and 1
hour would show as 00:00.000s.

After this change, durations are displayed, e.g., 1.234s, 1:23.456,
1:23:45.678.  No leading zero is displayed.  The 's' suffix is only
added if the duration is less than a minute.  Values greater than 24
hours will show with the full number of hours, e.g. 123:45:67.890.

This fixes #2726.