Use consistent field to return S3 HTTP request information
To perform a direct-to-S3 upload, a client makes a series of HTTP
requests to various /file endpoints. The responses include information
for the client to make an HTTP request to send data to S3 using a
pre-signed URL.

For a multi-part (chunked) upload, the `POST /file` and `POST
/file/chunk` endpoints return the request information in the
`s3.request` field. In contrast, the `POST /file/completion` endpoint
returns the request information in a `s3FinalizeRequest` field.

To make `POST /file/completion` consistent with the earlier steps in the
multi-part direct-to-S3 upload workflow, this commit changes the
response to use the `s3.request` field instead of `s3FinalizeRequest`.