The DICOM reader throws an exception for some data.
This prevents it from working with some valid images that refer to
delayed data, even though without that reference that would work.  When
such a IOError occurs, ignore it.

The pydicom module has some test files that are installed along with it.
Use one of these files to test that we can work with this data.