If a DICOM metadata has a binary value, it could prevent parsing.
We don't store values we don't understand, but we were trying to store
arbitrary binary values.  This fails if we can't coerce the value to
UTF-8 to stick it in Mongo.  If we actually want to store binary data,
we need to encode it different.  This skips such binary data, including
any data that doesn't encode to UTF-8 or has a null in it.

This uses a pydicom test file for testing.