Add remote_worker setup.py to the cache checksum
This is not ideal.  For one, the cache should be invalidated if any
plugin setup.py changes.  Second, we should not be caching the
virtualenv's containing editable installs.

The issue this fixes is related to
    https://github.com/pypa/pip/issues/4537

The downgrade of celery from #2761 triggers a behavior (maybe a bug?) in
pip that only occurs when the "egg-info" directory for an editable
install does *not* exist.  A second invocation of `pip install` will
run successfully.

To reproduce this problem outside of CI, in a fresh virtualenv:
```
pip install -e . -e ./plugins/jobs -e ./plugins/remote_worker
pip install celery==4.2
rm -fr ./plugins/remote_worker/*.egg-info
pip install -e ./plugins/remote_worker
```

One alternative option is to turn off caching of the virtualenv.