Improve the browser widget in the assetstore import.
Before, if you selected a user or collection as the destination for a
filesystem or S3 assetstore import, the ID was placed in the appropriate
field, but the type of resource was not changed.  This properly changes
the resource to folder, user, or collection.

This also displays the resource path next to the destination ID to let
the user know what they have selected.  Anything after the first
whitespace in the ID field is ignored, so this is for information
purposes only and has no effect on the import.