Sort imports when importing into a filesystem assetstore.
Sort files and directories when importing them into a filesystem
assetstore.  Try to import files first, as if we are importing to a
collection or user this will throw an exception right away instead of
possibly importing some of the directories and then failing.

By having a predictable sort order, if the task is interrupted due to a
failure or restart it is more obvious how far it got.