Add a setting to disable public access.
Note that plugins with @access.public endpoints may be broken if public
access is disabled.  For those endpoints, it may be necessary to change
to `@access.public(alwaysPublic=True)`.

The pytest access have been refactored to ensure that they use a clean
database, as the setting for public access now affects endpoints marked
public.