Scripting mode for girder-shell
This allows a script to be passed into the `girder-shell` or
`girder shell` command which will execute with the same context
as though its code were run inside the interactive shell. Namely,
its `__name__` will be `'__main__'`, it will have the `appconf`
and `webroot` globals set, and the necessary plugins loaded. Its
`sys.argv` will be whatever was passed after `--` of the `girder-shell`
invocation, e.g.

    girder shell --plugins=audit_logs ./audit_report.py -- --format=csv