Add loaded plugins to GET /system/plugins
Prior to this change, the enabled plugin list returned by this endpoint
was actually the list of currently loaded plugins.  Now, we return the
enabled list directly from the database setting and add a new key
containing the loaded list.  As a benefit, this would allow clients to
distinguish between plugins that were explicitly enabled vs. those that
were implicitly enabled by a dependency relationship.