Change behavior of ModelImporter
Fixes #2377. This removes the ability of the ModelImporter to
infer the class of a model based on the string(s), instead requiring
each model to be explicitly registered to a string if it is needed
for model loading.

This also changes the API for registering models to the ModelImporter;
rather than registering the model instance, one now only registers the
model type. This is cleaner in that it can defer model instantiation,
and also makes the ModelImporter decoupled from the fact that models
are singletons.

This is a breaking change.