Fix the client build.
Some upstream package used in building the client has changed.  It now
seems necessary to use the same version of underscore in the girder
package.json as in the web client src package.json and to include
babel-polyfill as part of the build.

This also backports the changes from #2999 to allow the build to pass
with the latest moto version.