Fix a misleading exception.
When asking for a folder that doesn't exist, this threw an exception
while checking if the folder was a virtual folder.  Rather, it should
just return and let the parent method handle the issue.

This specifically occurred when trying to find items in a folder that
didn't exist via a rest call (item?folderId=<non existant id>).