Remove unused code from test
This code was actually preventing us from using the `mock-db`
mode when running the test suite because it invoked a model
constructor outside of a test context. Since the database
mocking happens in the `db` fixture, instantiating a model outside
a test context will break (in the sense that it will try to
connect to the real database rather than the mock one).