Fix an issue removing uninstalled assetstore types.
You could not delete an assetstore that is a type that is no longer
present in the system, specifically because it attempted to delete
untracked uploads.  If it couldn't reach the assestore (for instance, an
S3 instance whose credentials were no longer valid), you could delete
the assetstore because this step was deemed non-essential.  However, if
you install a plugin to add an assetstore type, add an assetstore, and
the uninstall the plugin, you could not then remove the assetstore.

This guards that condition, so you do not need to reinstall the plugin
just to remove the assetstore.  Since the untracked uploads weren't
essential for removing an assetstore, this doesn't result in any
difference in functional behavior.

A similar guard was added to the enumeration of untracked uploads which
is accessible from the system/uploads endpoints.