Add an item/position/:id endpoint.
Mostly as an aid to determine which page of data is needed from a `GET
item` endpoint call, this adds a `GET item/position/:id` that takes the
same parameters as `GET item`, plus an item ID.  If the item is in the
result set, the response is the zero-based position within the results.
If the item is not in the result set, it is the position it would be
immediately after if it were to be added such that it would be in the
result set.  Note that since we only pass a single field for sort, items
can have the same sort rank.  In this case, they are subsorted by _id,
which, since the sub-order is undefined, may not actually reflect the
true position.

This does not have tests yet.

This should be extended to the folder model.  It could easily be added
to other models, too.

The virtual-folders plugin will need to be adjusted to work with this.