Add girder-readme plugin
The girder-readme plugin is used to show README files directly in the
hierarchy widget when viewing a folder, if that folder contains an item
starting with `README`.
All file extensions are valid (`README`, `README.txt`, `README.md`,
`README.rst`, etc.), but they will all be rendered as Markdown for
simplicity.
Behavior is undefined for multiple README items and for multiple files
in the item README; it will simply pick arbitrarily from the available
options.