Accept limit=0 on listResources for python client
limit=0 is a valid argument when listing resources on the REST-api, which,
if given, will list all items. The python client does not accept 0 as a valid
limit due to its "falsyness", and will override it with the default page
size.

This commit fixes the python-client so that limit=0 can be sent to
functions that internally call listResource.