Fixing scrolling and sorting of items
Fix how tox installs pip packages.

We explicitly stated an install_command that specified the local
directory.  This is part of how tox does installations in any case.
With recent versions of pip, this now fails.

This also adjusts some installation requirements for the pydicom
package, as the most recent version of that package has dropped support
for Python 3.6 without properly specifying so.

fixing centering issues and updating scrolling

mend