Remove eslint-plugin-underscore.
This removes some lint checks.  The module eslint-plugin-underscore
hasn't been updated since 2016 and relies on a version of lodash that is
considered to have a critical security issue.  This stops using it; it
really should be replaced with a better module, but a superficial search
doesn't turn up a newer module.