Remove the esdoc package.
The esdoc package doesn't actually get used in our CI.  We have
instructions for using it which have been updated.

Since this isn't actually used in CI and there are some dependabot
warnings on it, this reduces our warnings.  It probably would be better
to switch to jsdoc.

Also, with the package updates, fix a linting complaint.