Be more flexible with s3 read-only assetstore prefixes.
When a read-only s3 assetstore is created which uses a prefix, we test
if the prefix is listable.  This was being done as <bucket>/<prefix>,
but a common ACL is to authorize <bucket>/<prefix>/* and, in this case,
asking to list <bucket>/<prefix> fails.  Prior to this PR, this could be
worked around by specify "<prefix>/*", but this seems unintuitive.  This
PR first tests <bucket>/<prefix> for listability.  If that fails, it
tests <bucket>/<prefix>/, which will succeed in this use case.