Experimental package updates.
This builds and tests on node 14 / npm 6.  It claims to install on npm
8, but doesn't (yet?) build under node 16.

To install on npm >= 7, it is necessary to add `--install-links` to the
`npm install` command.  Otherwise, plugins' node modules aren't
installed; with it, they are installed alongside the girder node modules
(whereas with npm < 7, they are installed next to each plugin).

Upgrading to webpack 3 doesn't change much, but does show off that we
failed to properly ask for plugins during one of the client tests.

Currently, node 16 reports build success, but then fails to run because
`webpack.DllReferencePlugin` seems to not locate packages on start.