Add the ability to filter jobs by handler.
Some users run jobs remotely and locally (e.g., with the appropriate
plugins, running something via girder_worker and running an assetstore
import that is exposed as a job so that it is cancelable).  It is useful
to tell if any remote jobs are unfinished, as that can be used to
determine whether the remote worker should be started or stopped.