Load plugin static content as directed by server
This obviates the client having knowledge of plugin dependency ordering, as well as the need for plugins to be installed in the same `node_modules` as the client app.

This also removes the usage of `girder build` altogether. We could reintroduce it later if we want to add auto-building of client code, but it's not strictly necessary.