Clear the stored user and token on a 401 error.
When the user is not logged in, the server returns a 401 on some
endpoints.  This triggers showing the login dialog IF the browser
doesn't have a current user.  If you have two web pages open and you log
out in one, however, the second will get 401 errors but NOT show the
login dialog because it still tracks a current user.  This leaves the UI
showing errors but not a clear please-log-in-again method.