Add an assetstore method to get file size
This adds a `getFileSize()` method on the assetstore, which allows
assetstores to update a file size, if needed. If the file size
changes, the `File` model will update it.

We plan to use this with a DICOMweb assetstore adapter so that we do
not need to compute the file size right away, but only when it is needed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>