Remove the gridfsrs tests
They are failing because we aren't getting a report on which node is
primary.  The replicaset tests were necessary back when girder was first
created, but are no longer particularly useful as mongo has largely
abstracted the work of these away so they no longer need to be handled
by the client.