Harden S3 imports.
S3 buckets can have folders and items with the same name.  We want to
reuse existing folders so when we import a second time we don't
duplicate data.  In an item already exists with the name of the folder,
append a / (this is what S3 does internally) and use that as a name.