Add a show-download setting that defaults to all
This setting is core.download_shown and can be "all" (default), "user",
"admin", or "none".  If not all, download controls are only shown for a
subset of users -- any logged in user, only those who are admins, or no
one.

This doesn't effect endpoints, just what is shown in the UI.