Redesign S3 assetstore finalization process
This moves the multipart upload finalization step from the web client to the server.

While working in this code, I also tuned the part size from 32MB to 64MB to have fewer parts, to reduce the overhead in the multipart upload process while still retaining a reasonable amount of resumability for modern network performance.