Allow the default girder client transform job scope to be changed
By default, if a girder client is created, it is created with `scope=[
TokenScope.DATA_READ, TokenScope.DATA_WRITE]` (`['core.data.read',
'core.data.write']`).  It can be desired to grant girder clients more
scope than this.  If the environment variable GIRDER_WORKER_JOB_GC_SCOPE
is specified, this is treated as a comma-separated list of scopes OR as
a JSON encoded list of scope strings.  These are ADDED to the default
scope.  That is, specifying
`GIRDER_WORKER_JOB_GC_SCOPE="jobs.rest.create_job"` will give the girder
client read, write, and job creation privileges.