Inline the favicon as part of the vite build
The browser is requesting the favicon on every route change, even when
we only change the fragment.  This is different from Girder-3, probably
because of something that babel did in the webpack build.  I tried a
variety of solutions to prevent constantly re-requesting the favicon,
but this was the only one I found that actually worked reliably.