Pin bcrypt until passlib is updated
bcrypt 5.x makes a change that explicitly warns about truncating long
passwords.  passlib tests for a bug of some previous version by testing
with a long password, which raises this exception.

We currently use passlib for bcrypt and for one-time-passwords (2fa).
For the use of bcrypt, the change is straightforward.  We can get rid of
our `_cryptContext` object and replace any uses of it (two in
model/user.py and one in the rest user.py).  This will resolve the
current issue with passlib, and won't affect current salts, etc.

The OTP requires a bit more of a refactor, but we can switch to pyotp.
I haven't investigated whether this can be done without any major
change.